requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.4.0
billiard==3.6.4.0
boto3==1.18.21
botocore==1.21.21
CacheControl==0.12.6
cached-property==1.5.2
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-admin-interface==0.16.3
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-phonenumber-field==5.1.0
django-sendgrid==1.0.1
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
importlib-metadata==4.6.0
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==4.6.6
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.19
Pillow==8.3.0
progress==1.5
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycountry==20.7.3
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.2
python-http-client==3.3.2
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
s3transfer==0.5.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.14.0
smtpapi==0.3.1
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
